requirements.txt
# ===========================
# Imtehaan AI Tutor Backend
# Clean, Modern, Conflict-Free
# Python 3.10+
# ===========================

fastapi==0.104.1
uvicorn[standard]==0.24.0

# Core AI / LLM stack
openai>=1.10.0,<2.0.0
langchain==1.0.8
langchain-openai>=0.1.0
langgraph==1.0.3

# Data and async tools
pydantic==2.12.4
pydantic-core==2.41.5
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
anyio>=3.7.1,<4.0.0
httpx==0.28.1

# Supabase Python SDK stack
supabase==2.24.0
postgrest==2.24.0
realtime==2.24.0
storage3==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging & utilities
python-json-logger==2.0.7
PyYAML==6.0.3
requests==2.32.5

# For agents, embeddings, analytics
tqdm==4.67.1
colorama==0.4.6
zstandard==0.25.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

